refactoring of distance based classifiers
